requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.23
gunicorn==23.0.0
packaging==25.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
